Refocus tutorial on diff-drive wheel modifier calibration
