+++ ros/ros2_jazzy_turtlebot3_diffdrive_pixi/config/diff_drive_calibrator_sim.yaml
@@ -1,0 +1,9 @@
+diff_drive_calibrator:
+  ros__parameters:
+    use_sim_time: true
+    input_topic: "/cmd_vel_raw"
+    output_topic: "/cmd_vel"
+    wheel_radius_multiplier: 1.0
+    wheel_separation_multiplier: 1.0
+    cmd_timeout: 0.5
+    publish_rate_hz: 30.0
