--- ros/ros2_jazzy_turtlebot3_diffdrive_pixi/config/diff_drive_tuner_replay.yaml
@@ -1,15 +1,0 @@
-diff_drive_tuner:
-  ros__parameters:
-    use_sim_time: true
-    input_topic: "/cmd_vel"
-    output_topic: "/cmd_vel_tuned"
-    linear_scale: 1.0
-    angular_scale: 1.0
-    linear_bias: 0.0
-    angular_bias: 0.0
-    max_linear_speed: 0.22
-    max_angular_speed: 2.84
-    linear_accel_limit: 0.4
-    angular_accel_limit: 1.2
-    cmd_timeout: 0.5
-    publish_rate_hz: 30.0
