+++ ros/ros2_jazzy_turtlebot3_diffdrive_pixi/pixi.toml
@@ -26,8 +26,8 @@ ros-jazzy-turtlebot3-simulations = "*"
 
 [tasks]
 sim = "bash -lc 'ros2 launch \"$PIXI_PROJECT_ROOT/launch/headless_turtlebot3_gazebo.launch.py\"'"
-tune-sim = "python ./scripts/diff_drive_tuner.py --ros-args --params-file ./config/diff_drive_tuner_sim.yaml"
-tune-replay = "python ./scripts/diff_drive_tuner.py --ros-args --params-file ./config/diff_drive_tuner_replay.yaml"
+calibrate-sim = "python ./scripts/diff_drive_calibrator.py --ros-args --params-file ./config/diff_drive_calibrator_sim.yaml"
+calibrate-replay = "python ./scripts/diff_drive_calibrator.py --ros-args --params-file ./config/diff_drive_calibrator_replay.yaml"
 teleop = "ros2 run teleop_twist_keyboard teleop_twist_keyboard --ros-args -r /cmd_vel:=/cmd_vel_raw"
 drive-circle = "ros2 topic pub --rate 10 /cmd_vel_raw geometry_msgs/msg/Twist '{linear: {x: 0.2}, angular: {z: 0.4}}'"
 bag-record = "bash -lc 'mkdir -p bags && ros2 bag record -s mcap -o bags/tb3_$(date +%Y%m%d_%H%M%S) /clock /cmd_vel_raw /cmd_vel /odom /tf /tf_static /joint_states /scan'"
